backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
httpx==0.24.0
redis==4.5.0
python-dotenv==1.0.0
pytest==7.2.0
mangum==0.19.0